feat: better landing page
