+++ src/components/landing/LandingPageOneToOne.module.css
@@ -1,13 +1,12 @@
 .root {
-  /* Older layout, newer landing palette */
-  --lp-accent: #f8b84e;
+  --lp-accent: #f7ba57;
   --lp-orange: var(--brand-orange);
   --lp-blue: var(--brand-blue);
   --lp-teal: var(--teal);
   --lp-navy: var(--navy);
-  --lp-bg-0: #f7fbfa;
-  --lp-bg-1: #eef5f7;
-  --lp-bg-2: #ffffff;
+  --lp-bg-0: #edf4f7;
+  --lp-bg-1: #e5eef3;
+  --lp-bg-2: #f3f8fb;
   --lp-muted: #5f6f7f;
   --lp-muted-soft: #7c8a99;
 
@@ -66,17 +65,37 @@
 .nav {
   position: fixed;
   top: 0;
+  left: 0;
   width: 100%;
   z-index: 50;
-  background: rgba(255, 255, 255, 0.82);
+  background: rgba(245, 249, 252, 0.88);
   backdrop-filter: saturate(180%) blur(18px);
-  transition: all 0.3s ease;
+  transition:
+    top 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    width 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    left 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    transform 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    border-radius 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    background 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    box-shadow 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    border-color 0.48s cubic-bezier(0.16, 1, 0.3, 1),
+    backdrop-filter 0.48s cubic-bezier(0.16, 1, 0.3, 1);
   box-shadow: 0 1px 0 rgba(11, 17, 32, 0.06);
 }
 
 .scrolled {
-  background: rgba(255, 255, 255, 0.94);
-  box-shadow: 0 18px 48px rgba(11, 17, 32, 0.08);
+  top: 14px;
+  left: 50%;
+  width: min(calc(100% - 28px), 1440px);
+  transform: translateX(-50%);
+  border-radius: 999px;
+  border: 1px solid rgba(255, 255, 255, 0.58);
+  background: rgba(244, 249, 252, 0.62);
+  backdrop-filter: saturate(210%) blur(30px);
+  box-shadow:
+    0 28px 72px rgba(11, 17, 32, 0.12),
+    0 10px 24px rgba(255, 255, 255, 0.12) inset,
+    0 1px 0 rgba(255, 255, 255, 0.65) inset;
 }
 
 .navLink {
@@ -746,6 +765,12 @@
 }
 
 @media (max-width: 768px) {
+  .scrolled {
+    top: 10px;
+    width: calc(100% - 16px);
+    border-radius: 999px;
+  }
+
   .hamburger {
     display: flex;
   }
