+++ src/components/ui/aurora-background.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { cn } from "@/lib/utils";
 import React, { ReactNode } from "react";
 
@@ -67,7 +66,7 @@ export const AuroraBackground = ({
         </div>
       </main>
 
-      <style jsx global>{`
+      <style>{`
         @keyframes aurora {
           from {
             background-position: 50% 50%, 50% 50%;
